+++ src/app/debug/session-info/page.tsx
@@ -3,15 +3,15 @@
 import { useState } from "react";
 import { useSession } from "next-auth/react";
 import {
+  Badge,
+  Button,
   Card,
   CardContent,
   CardDescription,
   CardHeader,
   CardTitle,
-} from "@/components/ui/card";
-import { Badge } from "@/components/ui/badge";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
+  Input,
+} from "@/components/ui";
 import {
   RefreshCw,
   Download,
