+++ src/components/bulletin/ReplyForm.tsx
@@ -5,8 +5,7 @@
 "use client";
 
 import { useState } from "react";
-import { Button } from "@/components/ui/button";
-import { Avatar } from "@/components/ui/avatar";
+import { Avatar, Button } from "@/components/ui";
 import { Loader2 } from "lucide-react";
 import { useAuth } from "@/contexts/AuthContext";
